Made it clear that which class is to be used and supported and the obsolete classes (that may be still in use) inherit from the correct taghelper.
