+++ source/DasBlog.Web.Repositories/ArchiveManager.cs
@@ -12,7 +12,6 @@ namespace DasBlog.Managers
 	public class ArchiveManager : IArchiveManager
     {
         private readonly IBlogDataService dataService;
-        private readonly ILoggingDataService loggingDataService;
         private readonly IDasBlogSettings dasBlogSettings;
 
         public ArchiveManager(IDasBlogSettings settings)
