+++ source/DasBlog.Web.UI/TagHelpers/ApproveCommentTagHelper.cs
@@ -1,43 +1,14 @@
 ﻿using DasBlog.Services;
-using Microsoft.AspNetCore.Razor.TagHelpers;
+using DasBlog.Web.TagHelpers.Comments;
 using System;
-using System.Threading.Tasks;
 
 namespace DasBlog.Web.TagHelpers
 {
 	[Obsolete]
-    public class ApproveCommentTagHelper : TagHelper
-    {
-		public string BlogPostId { get; set; }
-
-		public string CommentId { get; set; }
-
-		public string CommentorName { get; set; }
-
-		private readonly IDasBlogSettings dasBlogSettings;
-		private const string COMMENTAPPROVE_URL = "{0}/comments/{1}";
-		private const string COMMENTTEXT_MSG = "Are you sure you want to approve the comment from '{0}'?";
-
-		public ApproveCommentTagHelper(IDasBlogSettings dasBlogSettings)
+    public class ApproveCommentTagHelper : CommentApprovalLinkTagHelper
+	{
+		public ApproveCommentTagHelper(IDasBlogSettings dasBlogSettings) : base(dasBlogSettings)
 		{
-			this.dasBlogSettings = dasBlogSettings;
-		}
-
-		public override void Process(TagHelperContext context, TagHelperOutput output)
-		{
-			var approvalurl = string.Format(COMMENTAPPROVE_URL, dasBlogSettings.GetPermaLinkUrl(BlogPostId), CommentId);
-			var commenttxt = string.Format(COMMENTTEXT_MSG, CommentorName);
-
-			output.TagName = "a";
-			output.TagMode = TagMode.StartTagAndEndTag;
-			output.Attributes.SetAttribute("href", $"javascript:commentManagement(\"{approvalurl}\",\"{commenttxt}\",\"PATCH\")");
-			output.Attributes.SetAttribute("class", "dbc-comment-approve-link");
-			output.Content.SetHtmlContent("Approve this comment");
-		}
-
-		public override Task ProcessAsync(TagHelperContext context, TagHelperOutput output)
-		{
-			return Task.Run(() => Process(context, output));
 		}
 	}
 }
