+++ source/DasBlog.Web.UI/TagHelpers/EditPostTagHelper.cs
@@ -1,40 +1,15 @@
 ﻿using DasBlog.Services;
-using DasBlog.Web.Models.BlogViewModels;
-using Microsoft.AspNetCore.Razor.TagHelpers;
+using DasBlog.Web.TagHelpers.Post;
 using System;
-using System.Threading.Tasks;
 
 namespace DasBlog.Web.TagHelpers
 {
 	[Obsolete]
-	public class EditPostTagHelper : TagHelper
+	public class EditPostTagHelper : PostEditLinkTagHelper
 	{
-		public PostViewModel Post { get; set; }
-		public string BlogPostId { get; set; }
-
-		private readonly IDasBlogSettings dasBlogSettings;
-
-		public EditPostTagHelper(IDasBlogSettings dasBlogSettings)
+		public EditPostTagHelper(IDasBlogSettings dasBlogSettings) : base(dasBlogSettings)
 		{
-			this.dasBlogSettings = dasBlogSettings;
-		}
 
-		public override void Process(TagHelperContext context, TagHelperOutput output)
-		{
-			if (Post != null)
-			{
-				BlogPostId = Post.EntryId;
-			}
-
-			output.TagName = "a";
-			output.TagMode = TagMode.StartTagAndEndTag;
-			output.Attributes.SetAttribute("href", dasBlogSettings.GetPermaLinkUrl(BlogPostId + "/edit"));
-			output.Content.SetHtmlContent("Edit this post");
-		}
-
-		public override Task ProcessAsync(TagHelperContext context, TagHelperOutput output)
-		{
-			return Task.Run(() => Process(context, output));
 		}
 	}
 }
