+++ source/DasBlog.Web.UI/TagHelpers/TitleLinkTagHelper.cs
@@ -1,5 +1,6 @@
 ﻿using DasBlog.Services;
 using DasBlog.Web.Models.BlogViewModels;
+using DasBlog.Web.TagHelpers.Post;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 using System;
 using System.Threading.Tasks;
@@ -7,39 +8,10 @@ using System.Threading.Tasks;
 namespace DasBlog.Web.TagHelpers
 {
 	[Obsolete]
-	public class TitleLinkTagHelper : TagHelper
+	public class TitleLinkTagHelper : PostTitleLinkTagHelper
 	{
-		public PostViewModel Post { get; set; }
-
-		public string Css { get; set; }
-
-		private readonly IDasBlogSettings dasBlogSettings;
-
-		public TitleLinkTagHelper(IDasBlogSettings dasBlogSettings)
+		public TitleLinkTagHelper(IDasBlogSettings dasBlogSettings) : base(dasBlogSettings)
 		{
-			this.dasBlogSettings = dasBlogSettings;
-		}
-
-		public override async Task ProcessAsync(TagHelperContext context, TagHelperOutput output)
-		{
-			output.TagMode = TagMode.StartTagAndEndTag;
-			output.TagName = "a";
-
-			var content = await output.GetChildContentAsync();
-			var title = content.GetContent();
-
-			if (!string.IsNullOrEmpty(Css))
-			{
-				output.Attributes.SetAttribute("class", Css);
-			}
-
-			if (string.IsNullOrWhiteSpace(title))
-			{
-				title = Post.Title;			
-			}
-
-			output.Attributes.SetAttribute("href", dasBlogSettings.RelativeToRoot(Post.PermaLink));
-			output.Content.SetHtmlContent(title);
 		}
 	}
 }
