--- source/DasBlog.Web.UI/TagHelpers/UserLinkTagHelper.cs
@@ -1,24 +1,0 @@
-﻿using Microsoft.AspNetCore.Razor.TagHelpers;
-using System;
-using System.Threading.Tasks;
-
-namespace DasBlog.Web.TagHelpers
-{
-	[Obsolete]
-	public class UserLinkTagHelper : TagHelper
-	{
-		public string EmailAddress { get; set; }
-		public string Linkability { get; set; }
-		public override void Process(TagHelperContext context, TagHelperOutput output)
-		{
-			output.TagName = "a";
-			output.TagMode = TagMode.StartTagAndEndTag;
-			output.Attributes.SetAttribute("href", $"javascript:linkToUser(\"{EmailAddress}\",\"{Linkability}\")");
-		}
-
-		public override Task ProcessAsync(TagHelperContext context, TagHelperOutput output)
-		{
-			return Task.Run(() => Process(context, output));
-		}
-	}
-}
